--- src/main/java/com/ideality/coreflow/notification/command/application/dto/NotificationDTO.java
@@ -1,16 +1,0 @@
-package com.ideality.coreflow.notification.command.application.dto;
-
-import lombok.*;
-
-// 알림 객체를 감싸는 클래스
-@Getter
-@Setter
-@ToString
-@Builder
-@RequiredArgsConstructor
-public class NotificationDTO{
-    private String message;
-    private String date;
-    private String status;
-
-}
